+++ src/main/java/org/fisco/bcos/web3j/protocol/core/JsonRpc2_0Web3j.java
@@ -23,6 +23,7 @@ import org.fisco.bcos.web3j.protocol.core.methods.response.BcosTransaction;
 import org.fisco.bcos.web3j.protocol.core.methods.response.BcosTransactionReceipt;
 import org.fisco.bcos.web3j.protocol.core.methods.response.BlockHash;
 import org.fisco.bcos.web3j.protocol.core.methods.response.BlockNumber;
+import org.fisco.bcos.web3j.protocol.core.methods.response.BlockTransactionReceipts;
 import org.fisco.bcos.web3j.protocol.core.methods.response.Call;
 import org.fisco.bcos.web3j.protocol.core.methods.response.Code;
 import org.fisco.bcos.web3j.protocol.core.methods.response.ConsensusStatus;
@@ -355,6 +356,44 @@ public class JsonRpc2_0Web3j implements Web3j {
                 Arrays.asList(groupId, transactionHash),
                 web3jService,
                 TransactionReceiptWithProof.class);
+    }
+
+    @Override
+    public Request<?, BlockTransactionReceipts> getBlockTransactionReceipts(
+            BigInteger blockNumber) {
+        return getBlockTransactionReceipts(blockNumber, BigInteger.ZERO, BigInteger.valueOf(-1));
+    }
+
+    @Override
+    public Request<?, BlockTransactionReceipts> getBlockTransactionReceipts(
+            BigInteger blockNumber, BigInteger offset, BigInteger count) {
+        return new Request<>(
+                "getBatchReceiptsByBlockNumberAndRange",
+                Arrays.asList(
+                        groupId,
+                        Numeric.encodeQuantity(blockNumber),
+                        offset.toString(10),
+                        count.toString(10),
+                        true),
+                web3jService,
+                BlockTransactionReceipts.class);
+    }
+
+    @Override
+    public Request<?, BlockTransactionReceipts> getBlockTransactionReceiptsByHash(
+            String blockHash) {
+        return getBlockTransactionReceiptsByHash(
+                blockHash, BigInteger.ZERO, BigInteger.valueOf(-1));
+    }
+
+    @Override
+    public Request<?, BlockTransactionReceipts> getBlockTransactionReceiptsByHash(
+            String blockHash, BigInteger offset, BigInteger count) {
+        return new Request<>(
+                "getBatchReceiptsByBlockHashAndRange",
+                Arrays.asList(groupId, blockHash, offset.toString(10), count.toString(10), true),
+                web3jService,
+                BlockTransactionReceipts.class);
     }
 
     @Override