+++ Changelog.md
@@ -1,3 +1,21 @@
+### v2.6.3
+
+(2020-12-23)
+
+* 新增
+1. getBlockTransactionReceipts
+2. getBlockTransactionReceiptsByHash 
+
+* 更新
+1. `netty-sm-ssl-context`升级至`1.2.0`
+
+* 修复
+1. `netty-all 4.1.52+`版本无法使用的问题
+2. 修复`jdk15`无法使用的问题
+
+* 兼容
+1. 适配fisco-bcos 2.4.0版本，支持Channel Message v1/V2/V3协议
+
 ### v2.6.2
 
 (2020-08-24)